use the Check Run outcome instead of the GHA API outcome
